# Implement 'faster on x% of runs'